Pick the most common footnote type in a text file.
This avoids having conflicting types of footnotes, for instance if an
anchor is at the beginning of a block (which can happen), while the
footnotes are marked as, e.g., [Note: ...].

The caller can also pass its own list of regexes to identify a
footnote. This may be used later in the interface.